将calarand设置界面移动回主界面
- 将calarand设置界面移动回主界面

已知问题：
- 修改部分设置时不会自动重置抽取器
